requirements.txt
fastapi==0.115.14
uvicorn==0.32.1
httpx==0.28.1
pydantic==2.11.0
pymongo==4.13.2
python-dateutil==2.9.0
python-dotenv==1.0.1
PyJWT==2.8.0